+++ src/pages/ui-elements/badge/Badge.js
@@ -40,14 +40,14 @@ const Badges = () => (
             Use the <code>pill</code> property to make badges more rounded
             (with a larger border-radius and additional horizontal padding).
           </p>
-          <Badge className="mr-xs" color="primary" pill>Primary</Badge>
-          <Badge className="mr-xs" color="info" pill>Info</Badge>
-          <Badge className="mr-xs" color="warning" pill>Warning</Badge>
-          <Badge className="mr-xs" color="success" pill>Success</Badge>
-          <Badge className="mr-xs" color="danger" pill>Danger</Badge>
-          <Badge className="mr-xs" color="secondary" pill>Secondary</Badge>
-          <Badge className="mr-xs" color="light" pill>Light</Badge>
-          <Badge className="mr-xs" color="dark" pill>Dark</Badge>
+          <Badge className="mr-md mb-md" color="primary" pill>Primary</Badge>
+          <Badge className="mr-md mb-md" color="info" pill>Info</Badge>
+          <Badge className="mr-md mb-md" color="warning" pill>Warning</Badge>
+          <Badge className="mr-md mb-md" color="success" pill>Success</Badge>
+          <Badge className="mr-md mb-md" color="danger" pill>Danger</Badge>
+          <Badge className="mr-md mb-md" color="secondary" pill>Secondary</Badge>
+          <Badge className="mr-md mb-md" color="light" pill>Light</Badge>
+          <Badge className="mr-md mb-md" color="dark" pill>Dark</Badge>
         </Widget>
       </Col>
       <Col xs={12} md={8}>
@@ -59,14 +59,14 @@ const Badges = () => (
             Using the contextual <code>href=&quot;&#35;&quot;</code> classes on
             an <code>&lt;Badge&gt;</code> element quickly provide actionable badges with hover and focus states.
           </p>
-          <Badge className="mr-xs" href="#" color="primary">Primary</Badge>
-          <Badge className="mr-xs" href="#" color="info">Info</Badge>
-          <Badge className="mr-xs" href="#" color="warning">Warning</Badge>
-          <Badge className="mr-xs" href="#" color="success">Success</Badge>
-          <Badge className="mr-xs" href="#" color="danger">Danger</Badge>
-          <Badge className="mr-xs" href="#" color="secondary">Secondary</Badge>
-          <Badge className="mr-xs" href="#" color="light">Light</Badge>
-          <Badge className="mr-xs" href="#" color="dark">Dark</Badge>
+          <Badge className="mr-md mb-md" href="#" color="primary">Primary</Badge>
+          <Badge className="mr-md mb-md" href="#" color="info">Info</Badge>
+          <Badge className="mr-md mb-md" href="#" color="warning">Warning</Badge>
+          <Badge className="mr-md mb-md" href="#" color="success">Success</Badge>
+          <Badge className="mr-md mb-md" href="#" color="danger">Danger</Badge>
+          <Badge className="mr-md mb-md" href="#" color="secondary">Secondary</Badge>
+          <Badge className="mr-md mb-md" href="#" color="light">Light</Badge>
+          <Badge className="mr-md mb-md" href="#" color="dark">Dark</Badge>
         </Widget>
       </Col>
     </Row>